Move IDE support into separate command file
