--- components/lark-hir/src/scope.rs
@@ -1,1 +1,0 @@
-
